+++ microservices/BookingService/BookingService.ClientHttp/Interfaces/IBookingClient.cs
@@ -3,26 +3,38 @@
 namespace BookingService.ClientHttp.Interfaces
 {
     /// <summary>
-    /// Client HTTP per l'interazione con il Booking Service.
-    /// Permette a servizi esterni di gestire le prenotazioni.
+    /// Client HTTP per l'interazione con il servizio BookingService.
     /// </summary>
+    /// <remarks>
+    /// Questa interfaccia definisce le operazioni disponibili per comunicare
+    /// con le API del servizio di gestione prenotazioni, tipicamente utilizzata
+    /// da layer esterni.
+    /// </remarks>
     public interface IBookingClient
     {
         /// <summary>
-        /// Recupera tutte le prenotazioni effettuate.
+        /// Recupera la lista di tutte le prenotazioni dal servizio remoto.
         /// </summary>
+        /// Una collezione di <see cref="BookingDto"/> oppure <c>null</c> in caso di errore.
         Task<IEnumerable<BookingDto>?> GetBookingsAsync();
 
         /// <summary>
-        /// Recupera il dettaglio di una singola prenotazione.
+        /// Recupera i dettagli di una prenotazione tramite identificativo.
         /// </summary>
+        /// <param name="id">Identificativo univoco della prenotazione.</param>
+        /// <returns>
+        /// Il <see cref="BookingDto"/> corrispondente oppure <c>null</c> se non trovato.
+        /// </returns>
         Task<BookingDto?> GetBookingByIdAsync(short id);
 
         /// <summary>
         /// Crea una nuova prenotazione nel sistema.
         /// </summary>
-        /// <returns>L'ID della prenotazione creata o null in caso di errore.</returns>
-        Task<int?> CreateBookingAsync(CreateBookingDto dto);
+        /// <param name="dto">Dati necessari alla creazione della prenotazione.</param>
+        /// <returns>
+        /// Il <see cref="CreateBookingDto"/> restituito dal servizio oppure <c>null</c> in caso di errore.
+        /// </returns>
+        Task<short?> CreateBookingAsync(CreateBookingDto dto);
 
         /// <summary>
         /// Aggiorna lo stato di una prenotazione (es. per completarla o annullarla).
